lz4_Frame_format.md: document kernel legacy last block
Since https://github.com/torvalds/linux/commit/2c484419efc09e7234c667aa72698cb79ba8d8ed, the Linux Kernel is able to discern the end of a legacy LZ4 frame by the presence of four 0 bytes, parsed as a B.CSize of 0 in the decoder. This is originally intended to make it mix well with grub initrd concatenation behavior (which uses a clause in the initramfs format about allowing arbitrary zero paddings), but I think it is a pretty good idea for any application limited to the legacy format in general.